fix move closing suggests
