+++ frontend/src/components/SearchSuggest/type.ts
